FIX(CPP02, ex01): add typecasting to float in toFloat() for its return value

+++ CPP_02/ex01/src/Fixed.cpp
@@ -60,7 +60,7 @@ void Fixed::setRawBits( int const raw )
 
 float Fixed::toFloat( void ) const
 {
-	return (_fix_point_value / (1 << _fractional_bits));
+	return (static_cast<float>(_fix_point_value) / (1 << _fractional_bits));
 }
 
 int Fixed::toInt( void ) const
